requirements.txt
fastapi
uvicorn[standard]
anthropic
elevenlabs
python-dotenv
python-multipart
httpx
beautifulsoup4
